+++ capture/afpacket/afring/afring.go
@@ -274,9 +274,9 @@ func (s *Source) Stats() (capture.Stats, error) {
 		return capture.Stats{}, err
 	}
 	return capture.Stats{
-		PacketsReceived: int(ss.Packets),
-		PacketsDropped:  int(ss.Drops),
-		QueueFreezes:    int(ss.QueueFreezes),
+		PacketsReceived: uint64(ss.Packets),
+		PacketsDropped:  uint64(ss.Drops),
+		QueueFreezes:    uint64(ss.QueueFreezes),
 	}, nil
 }
